[EliminationRound] persisting data to localStorage
