+++ .env
@@ -3,4 +3,4 @@ WS_PORT=5555
 
 # Configure Track Parameters
 LANE_COUNT=6
-TRACK_LENGTH_IN_FEET=10
+TRACK_LENGTH_IN_FEET=30
